fix: timer fixed and upgraded
Timer now shows up when the app is start but it starts counting after
the first tile is clicked. Hidding functionality is redone: It now
checks for the timer to reach 10 seconds instead of waiting for 10
seconds after the function call. It also shows up if the counter is less
than 10. The game reset button also stops and resets the counter back to
0.